# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/Ad/makeAd": {
            "post": {
                "tags": [
                    "Ad"
                ],
                "summary": "create new advertisement",
                "description": "",
                "operationId": "1",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Ad/Accept": {
            "put": {
                "tags": [
                    "Ad"
                ],
                "summary": "Accepting advertisement",
                "description": "",
                "operationId": "2",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Ad/GetByUser/{User_id}": {
            "get": {
                "tags": [
                    "Ad"
                ],
                "summary": "Return List of Ads for specific User",
                "description": "",
                "operationId": "3",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/bookmarks": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Find Bookmark By id",
                "description": "",
                "operationId": "4",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/bookmarks/add": {
            "post": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Add new Bookmark",
                "description": "",
                "operationId": "5",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/bookmarks/del": {
            "delete": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Delete Bookmark",
                "description": "",
                "operationId": "6",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Company/Get": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "company index",
                "description": "",
                "operationId": "7",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Company/Add": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Add company",
                "description": "",
                "operationId": "8",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Company/Get/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Find Company By id",
                "description": "",
                "operationId": "9",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Company/Update/{id}": {
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update company",
                "description": "",
                "operationId": "10",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Company/Delete/{id}": {
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete company",
                "description": "",
                "operationId": "11",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Request/Add": {
            "post": {
                "tags": [
                    "Request"
                ],
                "summary": "create a new request",
                "description": "",
                "operationId": "12",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Request/Accept": {
            "put": {
                "tags": [
                    "Request"
                ],
                "summary": "Accept the fucking request and Reject other fucking requests",
                "description": "",
                "operationId": "13",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Request/Reject": {
            "put": {
                "tags": [
                    "Request"
                ],
                "summary": "Reject the fucking request",
                "description": "",
                "operationId": "14",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Request/GetByUser/{User_id}": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Return List of Requests For User's Ads",
                "description": "",
                "operationId": "15",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/Request/GetByAd/{Ad_id}": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Return List of Requests For specific Ad",
                "description": "",
                "operationId": "16",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user",
                "description": "",
                "operationId": "17",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user",
                "description": "",
                "operationId": "18",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logout from account",
                "description": "",
                "operationId": "19",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}